Move `PageTemplate` into `src/site`
