+++ src/site/PageTemplate.js
@@ -1,0 +1,36 @@
+import { Box, VStack } from "@chakra-ui/react";
+import Card from "../common/components/Card";
+import Navbar from "../common/components/navbar/Navbar";
+
+/**
+ * Template component for a page with the sidebar included.
+ *
+ * The children elements of the page will be vertically stacked.
+ * A typical body of a page includes a breadcrumb trail before the rest of the content.
+ *
+ * @param heading the heading only used for the authentication pages
+ * @param children the children elements of the page
+ * @param buttons an array of components to be displayed in the navbar
+ * @return the page populated with the sidebar and children elements
+ */
+export default function PageTemplate({ heading, children, buttons = [] }) {
+    return (
+        <VStack>
+            <Navbar buttons={buttons} />
+            <VStack
+                pt={[20, 24]}
+                justify="center"
+                spacing="0"
+                w="100%"
+                minH="100vh"
+                overflowX="hidden"
+            >
+                <Box w={["100vw", "100vw", "100vw", "80vw", "60vw"]}>
+                    <Card isCentered isStandalone heading={heading}>
+                        {children}
+                    </Card>
+                </Box>
+            </VStack>
+        </VStack>
+    );
+}
