+++ src/site/errors/WorkInProgress.js
@@ -1,5 +1,5 @@
 import Breadcrumbs from "../../common/components/Breadcrumbs";
-import PageTemplate from "../../common/components/PageTemplate";
+import PageTemplate from "../PageTemplate";
 import { Box, Button, Heading, Text, VStack } from "@chakra-ui/react";
 import { useNavigate } from "react-router-dom";
